+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 
 - Generator
   - Updated JSON schema arrays to use `contains` when applicable.
+- Schema
+  - Added support for `regex` types (behind the `type_regex` feature).
 
 #### 🐞 Fixes